pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "alertivo"
version = "0.1.0"
description = "Alertivo desktop monitoring app with local control center, tray mode and simple onboarding."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiohttp>=3.11.0",
  "beautifulsoup4>=4.12.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]
desktop = [
  "pywebview>=5.0",
  "pystray>=0.19.5",
  "Pillow>=10.4.0",
]

[project.scripts]
alertivo-monitor = "gpu_alerts.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-s"

[tool.hatch.build.targets.wheel]
packages = ["src/gpu_alerts"]
include = ["src/gpu_alerts/control_center_template.html"]
